replaced order logic into OrderView
